+++ rollup.config.mjs
@@ -1,7 +1,10 @@
+/*  eslint-env node */
 import babel from "rollup-plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
 import resolve from "@rollup/plugin-node-resolve";
-import rollupPluginTerser from "rollup-plugin-terser";
+import terser from "rollup-plugin-terser";
+
+const IS_DEV = process.env.ROLLUP_WATCH;
 
 const INPUT_NAME = "index.mjs";
 const OUTPUT_NAME = "of";
@@ -32,14 +35,15 @@ export default {
     babel({ babelrc: true }),
     resolve(),
     commonjs(),
-    rollupPluginTerser.terser({
-      keep_classnames: true,
-      keep_fnames: true,
-      output: {
-        comments: false,
-      },
-      sourcemap: true,
-      warnings: true,
-    }),
+    !IS_DEV &&
+      terser.terser({
+        keep_classnames: true,
+        keep_fnames: true,
+        output: {
+          comments: false,
+        },
+        sourcemap: true,
+        warnings: true,
+      }),
   ],
 };