+++ src/core/ofAny.mjs
@@ -1,5 +1,5 @@
-import { ERR_UNKNOWN } from "../const/error";
-import { FUNC, FUNC_ASYNC, FUNC_GEN, OBJ_GEN, PROMISE } from "../const/type";
+import { ERR_UNKNOWN } from "../const/error.mjs";
+import { FUNC, FUNC_ASYNC, FUNC_GEN, OBJ_GEN, PROMISE } from "../const/type.mjs";
 
 /**
  * @name ofAny