+++ src/core/ofSync.mjs
@@ -1,6 +1,6 @@
-import { ERR_UNKNOWN } from "../const/error";
+import { ERR_UNKNOWN } from "../const/error.mjs";
 
-import { isFunction } from "../util/isFunction";
+import { isFunction } from "../util/isFunction.mjs";
 
 /**
  * @name ofSync